+++ project/core/src/net/ashald/envfile/EnvFileErrorException.java
@@ -1,0 +1,16 @@
+package net.ashald.envfile;
+
+
+public class EnvFileErrorException extends Exception {
+    public EnvFileErrorException(String message) {
+        super(message);
+    }
+
+    public EnvFileErrorException(Throwable cause) {
+        super(cause);
+    }
+
+    public EnvFileErrorException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}